pyproject.toml
[project]
name = "zimage-cli"
version = "0.1.0"
description = "Command tool for z-image model"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "diffusers",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
]

[tool.uv.sources]
diffusers = { git = "https://github.com/huggingface/diffusers" }

